Made a (unfinished) web-app frontend and route to upload .csv files
I probably should have made two separate commits, but this is just my stopping point
for now. I made a quick web-app frontend with lovable AI and modified the routes.js
to handle file upload. The web-app runs fine, but need to connect the backend logic
to it so it is functional. Good starting point so far.

Design of webapp is up to changes, used lovable AI for a quick start.