Final commit message for Spring Semester 2025 This version of the web-app has little to no issues. It features: - Working submission of Mantis Data - Working fetching of Heart Rate Data from the watch - Working fetching of the data into the plots
It does not yet feature:
- "Key Insights" for the data. They are currently just sample insights.
- Does not populate the alternative view to the plots for the data.

Issues to address:
- Due to focusing on developing rather than production, only one user is supported.
The database itself supports multiple users, but we have not implemented the logic
to change between users (login/logout).
- When uploading the Mantis data, it actually flushes out all of the data and then
writes the new data. This is because the Mantis data only comes in one csv with ALL
the sessions. This is rather easily fixable by adding one more step to processing
the data (if sessionID is not in the database, add it, else terminate).
Fastest way to do this is working with the fact that the mantis data sessions
data comes sorted from oldest to most recent. Most recent being the last session in
the database (Important distinction: In the database, last session is most recent, but
in the csv, the FIRST session is the most recent. It is different because data is being
inserted one by one from the top of the csv to the bottom)
- Lastly, the heart rate data automatically inherits the sessionID from the
most recent mantis session. This should be chnaged so the logic is a bit smarter
and secure, but it works under the assumption that the user will upload the mantis data
first.

We were also going to change the website's design to be a bit more aesthetically pleasing,
so look into that as well.